Handle NoSuchMethodError
In https://www.github.com/swoval/swoval, the javafmt task wasn't working
because an old version of guava was in the classpath. I couldn't tell
what was going on since there was no error reporting. There is a
workaround to the error, but I'd like to figure out how to fix this long
term.