Use reflection to try to guarantee a sane guava version
The scala.js plugin brings in a fat jar that includes an older version
of guava. This causes the plugin to not work if it is listed after the
scala.js plugin in project/plugins.sbt. To work around this, I use a
custom ClassLoader to load the JavaFormatter companion object that
ensures that the guava jar is at the head of the url list.