Reimplement plugin
This commit reworks the plugin to use the new file tracking apis
introduced in sbt 1.3.0. The new creates an intermediate task that
computes all of the correctly formatted files. This intermediate task
is invoked dynamically so that we can toggle whether or not it
overwrites the source file if it is incorrectly formatted and it can be
formatted. By sharing the task, we can reuse the file stamp cache
between the check and format tasks to avoid spurious reformatting. I
also reworked things so that each format evaluation runs in a separate
task. This allows supershell to display which files are being formatted.
It also allows the formatting to run in parallel.